// src/entities/weather/types/weather.ts
export interface Weather {
  id: number
  name: string
  country: string
  iconUrl: string
  main: string
  temp: number
  feelsLike: number
  description: string
  windSpeed: number
  windDeg: number
  windDir: string
  humidity: number
  pressure: number
  seaLevel?: number
  visibility: number
}